requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
chromadb>=0.4.0
datasets>=2.14.0
pandas>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0
pydantic>=2.0.0
numpy>=1.24.0
tqdm>=4.66.0

